+++ packages/service-library/src/servicelib/redis/__init__.py
@@ -1,0 +1,14 @@
+from ._client import RabbitMQClient
+from ._errors import (
+    CouldNotAcquireLockError,
+    CouldNotConnectToRedisError,
+    LockLostError,
+)
+
+__all__: tuple[str, ...] = (
+    "CouldNotAcquireLockError",
+    "CouldNotConnectToRedisError",
+    "LockLostError",
+)
+
+# nopycln: file
